Seperate InstantSearch and Hits into seperate components
- Add available to hits
